+++ src/extractors/undefined-extractor.ts
@@ -2,14 +2,7 @@ import { BaseExtractor } from './base-extractor.js';
 
 export class UndefinedExtractor extends BaseExtractor {
 	protected _process(parameter: any): any {
-		if (
-			parameter.hasOwnProperty('name') &&
-			parameter.name === 'return' &&
-			Object.keys(parameter).length === 1
-		) {
-			return undefined;
-		}
-
+		// Always returns undefined to pass through to next extractor
 		return undefined;
 	}
 }
