+++ src/shared/results-utils.ts
@@ -1,0 +1,42 @@
+/**
+ * Compares two results for equality, handling nested objects and numbers
+ */
+export function resultsEqual(expected: any, actual: any): boolean {
+	if (expected === undefined && actual === undefined) {
+		return true;
+	}
+
+	if (expected === null && actual === null) {
+		return true;
+	}
+
+	if (typeof expected === 'number') {
+		return Math.abs(actual - expected) < 0.00000001;
+	}
+
+	if (expected === actual) {
+		return true;
+	}
+
+	if (
+		typeof expected !== 'object' ||
+		expected === null ||
+		typeof actual !== 'object' ||
+		actual === null
+	) {
+		return false;
+	}
+
+	const expectedKeys = Object.keys(expected);
+	const actualKeys = Object.keys(actual);
+
+	if (expectedKeys.length !== actualKeys.length) return false;
+
+	for (const key of expectedKeys) {
+		if (!actualKeys.includes(key) || !resultsEqual(expected[key], actual[key])) {
+			return false;
+		}
+	}
+
+	return true;
+}
